example_registry/pyproject.toml
[project]
name = "function-registry"
version = "0.1.0"
description = "FastAPI-based function call registry with auto-discovery"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]
web = [
    "firecrawl-py>=0.0.1",
    "tavily-python>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["functions", "models.py", "function_discovery.py", "main.py"]

